+++ scanapi/reporter.py
@@ -1,17 +1,14 @@
 #!/usr/bin/env python3
 import datetime
-import logging
 import pathlib
 import webbrowser
 
 from pkg_resources import get_distribution
 
+from scanapi.console import write_report_path
 from scanapi.session import session
 from scanapi.settings import settings
 from scanapi.template_render import render
-from scanapi.test_status import TestStatus
-
-logger = logging.getLogger(__name__)
 
 
 class Reporter:
@@ -28,7 +25,7 @@ class Reporter:
         self.output_path = pathlib.Path(output_path or "scanapi-report.html")
         self.template = template
 
-    def write(self, results):
+    def write(self, results, open_in_browser):
         """Part of the Reporter instance that is responsible for writing
         scanapi-report.html.
 
@@ -39,8 +36,6 @@ class Reporter:
             None
 
         """
-        logger.info("Writing documentation")
-
         template_path = self.template if self.template else "report.html"
         has_external_template = bool(self.template)
         context = self._build_context(results)
@@ -50,31 +45,14 @@ class Reporter:
         with open(self.output_path, "w", newline="\n") as doc:
             doc.write(content)
 
-        logger.info("\nThe documentation was generated successfully.")
-        logger.info(f"It is available at {self.output_path.resolve().as_uri()}")
+        write_report_path(self.output_path.resolve().as_uri())
 
-    def open_report_in_browser(self):
+        if open_in_browser:
+            self._open_in_browser()
+
+    def _open_in_browser(self):
         """Open the results file on a browser"""
         webbrowser.open(self.output_path.resolve().as_uri())
-
-    @staticmethod
-    def write_without_generating_report(results):
-        """Part of the Reporter instance that is responsible for writing the
-        results without generating the scanapi-report.html.
-
-        Args:
-            results [generator]: generator of dicts resulting of Request run().
-
-        Returns:
-            None
-        """
-        logger.info("Writing results without generating report")
-        for r in results:
-            if logger.root.level != logging.DEBUG:
-                for test in r["tests_results"]:
-                    logger.info(f" [{test['status'].upper()}] {test['name']}")
-                    if test["status"] == TestStatus.FAILED:
-                        logger.info(f"\t {test['failure']} is false")
 
     @staticmethod
     def _build_context(results):