+++ scanapi/tree/request_node.py
@@ -1,6 +1,6 @@
-import logging
 import time
 
+from scanapi.console import console
 from scanapi.errors import HTTPMethodNotAllowedError
 from scanapi.hide_utils import hide_sensitive_info
 from scanapi.test_status import TestStatus
@@ -18,8 +18,6 @@ from scanapi.tree.tree_keys import (
     VARS_KEY,
 )
 from scanapi.utils import join_urls, session_with_retry, validate_keys
-
-logger = logging.getLogger(__name__)
 
 
 class RequestNode:
@@ -125,7 +123,7 @@ class RequestNode:
 
         method = self.http_method
         url = self.full_url_path
-        logger.info("Making request %s %s", method, url)
+        console.print(f"\n- Making request {method} {url}", highlight=False)
 
         self.endpoint.spec_vars.update(
             self.spec.get(VARS_KEY, {}),