+++ tests/unit/test_scan.py
@@ -7,7 +7,7 @@ import yaml
 from pytest import fixture, mark, raises
 
 from scanapi.errors import EmptyConfigFileError, InvalidKeyError
-from scanapi.scan import open_report_in_browser, scan, write_report
+from scanapi.scan import scan, write_report
 
 log = logging.getLogger(__name__)
 
@@ -121,7 +121,7 @@ class TestScan:
         )
 
     @mark.context("when the api spec is ok")
-    @mark.it("should call reporter write_report")
+    @mark.it("should call reporter write_report and console write_results")
     def test_should_call_reporter(self, mocker, response):
         mock_load_config_file = mocker.patch("scanapi.scan.load_config_file")
         mock_load_config_file.return_value = {"endpoints": []}
@@ -130,6 +130,7 @@ class TestScan:
         mock_endpoint_run = mocker.patch("scanapi.scan.EndpointNode.run")
         mock_endpoint_run.return_value = [response]
         mock_write_report = mocker.patch("scanapi.scan.write_report")
+        mock_write_results = mocker.patch("scanapi.scan.write_results")
 
         with raises(SystemExit) as excinfo:
             scan()
@@ -139,12 +140,13 @@ class TestScan:
 
         mock_endpoint_init.assert_called_once_with({"endpoints": []})
         assert mock_endpoint_run.called
-        mock_write_report.assert_called_once_with([response])
+        mock_write_report.assert_called_once_with([response], False)
+        assert mock_write_results.called
 
     @mark.context(
         "when the api spec is ok and the is configured to open the results"
     )
-    @mark.it("should call reporter write_report and open_report_in_browser")
+    @mark.it("should call reporter write_report")
     def test_should_call_reporter_and_open_results(self, mocker, response):
         mock_load_config_file = mocker.patch("scanapi.scan.load_config_file")
         mock_load_config_file.return_value = {"endpoints": []}
@@ -153,9 +155,8 @@ class TestScan:
         mock_endpoint_run = mocker.patch("scanapi.scan.EndpointNode.run")
         mock_endpoint_run.return_value = [response]
         mock_write_report = mocker.patch("scanapi.scan.write_report")
-        mock_open_report_in_browser = mocker.patch(
-            "scanapi.scan.open_report_in_browser"
-        )
+        mock_write_results = mocker.patch("scanapi.scan.write_results")
+
         mocker.patch(
             "scanapi.scan.settings",
             {
@@ -176,14 +177,14 @@ class TestScan:
 
         mock_endpoint_init.assert_called_once_with({"endpoints": []})
         assert mock_endpoint_run.called
-        mock_write_report.assert_called_once_with([response])
-        mock_open_report_in_browser.assert_called_once()
+        assert mock_write_results.called
+        mock_write_report.assert_called_once_with([response], True)
 
 
 @mark.describe("scan")
 @mark.describe("write_report")
 class TestWriteReport:
-    @mark.it("should call wr")
+    @mark.it("should call write_report")
     def test_should_call_wr(self, mocker, response):
         mock_write = mocker.patch("scanapi.scan.Reporter.write")
         mock_reporter_init = mocker.patch("scanapi.scan.Reporter.__init__")
@@ -198,36 +199,9 @@ class TestWriteReport:
             },
         )
 
-        write_report([response])
+        write_report([response], False)
 
         mock_reporter_init.assert_called_once_with(
             "out/my-report.md", "my-template.jinja"
         )
-        mock_write.assert_called_once_with([response])
-
-
-@mark.describe("scan")
-@mark.describe("open_report_in_browser")
-class TestOpenReport:
-    @mark.it("should call open browser")
-    def test_should_call_wr(self, mocker, response):
-        mock_open = mocker.patch("scanapi.scan.Reporter.open_report_in_browser")
-        mock_reporter_init = mocker.patch("scanapi.scan.Reporter.__init__")
-        mock_reporter_init.return_value = None
-        mocker.patch(
-            "scanapi.scan.settings",
-            {
-                "output_path": "out/my-report.md",
-                "no_report": False,
-                "reporter": "markdown",
-                "template": "my-template.jinja",
-            },
-        )
-
-        open_report_in_browser()
-
-        mock_reporter_init.assert_called_once_with(
-            "out/my-report.md", "my-template.jinja"
-        )
-
-        mock_open.assert_called_once()
+        mock_write.assert_called_once_with([response], False)