+++ tests/unit/tree/testing_node/test_run.py
@@ -1,10 +1,6 @@
-import logging
-
 from pytest import fixture, mark
 
 from scanapi.tree import EndpointNode, RequestNode, TestingNode
-
-log = logging.getLogger(__name__)
 
 
 @mark.describe("testing node")
@@ -68,15 +64,6 @@ class TestRun:
         assert mock_increment_successes.call_count == 1
         assert not mock_increment_failures.called
 
-    @mark.context("when test passed")
-    @mark.it("should logs test results")
-    def test_logs_test_results(self, mock_evaluate, caplog, testing_node):
-        mock_evaluate.return_value = (True, None)
-
-        with caplog.at_level(logging.DEBUG):
-            testing_node.run()
-        assert "\x07 [PASSED] foo::bar::status_is_200" in caplog.text
-
     @mark.context("when test failed")
     @mark.it("should build result object")
     def test_build_result_2(
@@ -112,17 +99,3 @@ class TestRun:
         testing_node.run()
         assert mock_increment_failures.call_count == 1
         assert not mock_increment_successes.called
-
-    @mark.context("when test failed")
-    @mark.it("should logs test results")
-    def test_logs_test_results_2(self, mock_evaluate, caplog, testing_node):
-        mock_evaluate.return_value = (
-            False,
-            "response.status_code == 200",
-        )
-
-        with caplog.at_level(logging.DEBUG):
-            testing_node.run()
-
-        assert "\x07 [FAILED] foo::bar::status_is_200" in caplog.text
-        assert "\t  response.status_code == 200 is false" in caplog.text