+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Print test results summary in console [#497](https://github.com/scanapi/scanapi/issues/497)
 
 ## [2.6.1] - 2022-04-12
 ### Changed